Refs #24. Fix import paths, move assert_* helpers to util
